milestone

+++ paddle/fluid/platform/timer.cc
@@ -49,7 +49,7 @@ bool Timer::IsPaused() {
   return _paused;
 }
 
-int Timer::Count() { return _count; }
+long Timer::Count() { return _count; }
 
 double Timer::ElapsedUS() { return static_cast<double>(_elapsed); }
